Figured out how to do GCM auth check
